FIX: Refactor
